complete tweet service

+++ src/main/java/entities/Likes.java
@@ -6,14 +6,14 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Likes {
-    private int id;
-    private int tweet_id;
-    private int user_id;
+    private long id;
+    private long tweet_id;
+    private long user_id;
     private boolean isLike;
 
     public Likes() {
     }
-    public Likes(int tweet_id, int user_id, boolean isLike) {
+    public Likes(long tweet_id, long user_id, boolean isLike) {
         this.tweet_id = tweet_id;
         this.user_id = user_id;
         this.isLike = isLike;
