+++ src/main/java/repository/TagRepository.java
@@ -32,10 +32,10 @@ public class TagRepository {
         }
     }
 
-    public List<Tags> getTagsByTweetId(int tweetId) throws SQLException {
+    public List<Tags> getTagsByTweetId(long tweetId) throws SQLException {
         List<Tags> tags = new ArrayList<>();
         try (var statement = Datasource.getConnection().prepareStatement(READ_TAGS_BY_TWEET_ID)) {
-            statement.setInt(1, tweetId);
+            statement.setLong(1, tweetId);
             try (ResultSet resultSet = statement.executeQuery()) {
                 while (resultSet.next()) {
                     int id = resultSet.getInt("id");
